feat: add store/send text commands and refactoring
This commit adds the following:

* Command: /addText

This option takes the text of the message sent by the user and stores it
in a file in the res folder.

If file doesn't exists it's created. Otherwise, the text is appended at
the end.

* Command: /getText

In this case we get a random line of the file containing the text
stored by the user.

* Refactor BotActionsHandler.java

The code has been refactored following the Clean Code approach.